pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "xnat-exported-scans"
description = "Uploads exported DICOM and raw data to XNAT, parsing metadata from DICOMs"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "click >=8.1",
    "fileformats-medimage-extras",
    "pydicom >=2.3.1",
    "tqdm >=4.64.1",
    "xnat",
]
license = { file = "LICENSE" }
authors = [{ name = "Thomas G. Close", email = "thomas.close@sydney.edu.au" }]
maintainers = [
    { name = "Thomas G. Close", email = "thomas.close@sydney.edu.au" },
]
keywords = ["xnat", "imaging", "data", "repository", "siemens", "pet", "upload"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering",
]
dynamic = ["version"]

[project.optional-dependencies]
dev = ["black", "pre-commit", "codespell", "flake8", "flake8-pyproject"]
test = [
    "pytest >=6.2.5",
    "pytest-env>=0.6.2",
    "xnat4tests >=0.3",
    "medimages4tests >=0.3.1",
    "PyYAML",
]
# Aliases
tests = ["xnat-exported-scans[test]"]
all = ["xnat-exported-scans[dev,test]"]

[project.urls]
repository = "https://github.com/Australian-Imaging-Service/xnat_exported_scans"

[project.scripts]
xnat-ingest = "xnat_ingest.cli:cli"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "xnat_exported_scans/_version.py"

[tool.hatch.build]
packages = ["xnat_exported_scans"]
exclude = ["/tests"]

[tool.black]
target-version = ['py37', 'py38']
exclude = "xnat_exported_scans/_version.py"

[tool.codespell]
ignore-words = ".codespell-ignorewords"

[tool.flake8]
doctests = true
per-file-ignores = ["__init__.py:F401"]
max-line-length = 88
select = "C,E,F,W,B,B950"
extend-ignore = ['E203', 'E501', 'E129', "W503"]
